GH-2: Support user defined non-XSD data types
- for properties with rdf:range, CimProfileRegistry now checks the Jena TypeMapper for registered RDFDataTypes before treating the rdf:range value as reference
  - added test for this behaviour
- added test for supporting custom datatypes
- Refactored SPARQL properties query
- updated formatting of tests to new style conventions
